pyproject.toml
[project]
name = "googleadk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.25.0",
    "httpx>=0.28.1",
    "litellm>=1.81.11",
]
